refactor: Change the return type of GetPublicPort to uint instead of uint16
All references to ports in the codebase use uint not uint16.
There is no reason to keep the ports as they are

+++ cluster-provision/gocli/providers/start.go
@@ -63,7 +63,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 		if err != nil {
 			return err
 		}
-		kp.SSHPort = port
+		kp.SSHPort = uint(port)
 	}
 
 	if kp.APIServerPort == 0 {
@@ -71,7 +71,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 		if err != nil {
 			return err
 		}
-		kp.APIServerPort = port
+		kp.APIServerPort = uint(port)
 	}
 
 	registry, err := kp.runRegistry(ctx)
@@ -94,7 +94,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 
 	for x := 0; x < int(kp.Nodes); x++ {
 		nodeName := kp.nodeNameFromIndex(x + 1)
-		sshClient, err := libssh.NewSSHClient(kp.SSHPort, x+1, false)
+		sshClient, err := libssh.NewSSHClient(uint16(kp.SSHPort), x+1, false)
 		if err != nil {
 			return err
 		}
@@ -161,7 +161,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 		if err = rootkey.Exec(); err != nil {
 			return err
 		}
-		sshClient, err = libssh.NewSSHClient(kp.SSHPort, x+1, true)
+		sshClient, err = libssh.NewSSHClient(uint16(kp.SSHPort), x+1, true)
 
 		if err = kp.provisionNode(sshClient, x+1); err != nil {
 			return err
@@ -173,7 +173,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 		}(node.ID)
 	}
 
-	sshClient, err := libssh.NewSSHClient(kp.SSHPort, 1, true)
+	sshClient, err := libssh.NewSSHClient(uint16(kp.SSHPort), 1, true)
 	if err != nil {
 		return err
 	}
@@ -188,7 +188,7 @@ func (kp *KubevirtProvider) Start(ctx context.Context, cancel context.CancelFunc
 		return err
 	}
 
-	config, err := k8s.NewConfig(".kubeconfig", kp.APIServerPort)
+	config, err := k8s.NewConfig(".kubeconfig", uint16(kp.APIServerPort))
 	if err != nil {
 		return err
 	}
